Refactor console log streaming
